rename getters to deps
