+++ src/backend/langflow/__main__.py
@@ -11,7 +11,7 @@ import typer
 from dotenv import load_dotenv
 from langflow.main import setup_app
 from langflow.services.database.utils import session_getter
-from langflow.services.getters import get_db_service, get_settings_service
+from langflow.services.deps import get_db_service, get_settings_service
 from langflow.services.utils import initialize_services, initialize_settings_service
 from langflow.utils.logger import configure, logger
 from multiprocess import Process, cpu_count  # type: ignore
