+++ src/backend/langflow/api/v1/chat.py
@@ -15,10 +15,10 @@ from langflow.graph.graph.base import Graph
 from langflow.services.auth.utils import get_current_active_user, get_current_user
 from langflow.services.cache.utils import update_build_status
 from loguru import logger
-from langflow.services.getters import get_chat_service, get_session, get_cache_service
+from langflow.services.deps import get_chat_service, get_session, get_cache_service
 from sqlmodel import Session
-from langflow.services.chat.manager import ChatService
-from langflow.services.cache.manager import BaseCacheService
+from langflow.services.chat.service import ChatService
+from langflow.services.cache.service import BaseCacheService
 
 
 router = APIRouter(tags=["Chat"])
