+++ src/backend/langflow/api/v1/components.py
@@ -2,7 +2,7 @@ from datetime import timezone
 from typing import List
 from uuid import UUID
 from langflow.services.database.models.component import Component, ComponentModel
-from langflow.services.getters import get_session
+from langflow.services.deps import get_session
 from sqlmodel import Session, select
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.exc import IntegrityError
