+++ src/backend/langflow/interface/custom/custom_component.py
@@ -4,7 +4,7 @@ from fastapi import HTTPException
 from langflow.field_typing.constants import CUSTOM_COMPONENT_SUPPORTED_TYPES
 from langflow.interface.custom.component import Component
 from langflow.interface.custom.directory_reader import DirectoryReader
-from langflow.services.getters import get_db_service
+from langflow.services.deps import get_db_service
 from langflow.interface.custom.utils import extract_inner_type
 
 from langflow.utils import validate
