+++ src/backend/langflow/interface/listing.py
@@ -1,4 +1,4 @@
-from langflow.services.getters import get_settings_service
+from langflow.services.deps import get_settings_service
 from langflow.utils.lazy_load import LazyLoadDictBase
