+++ src/backend/langflow/interface/output_parsers/base.py
@@ -4,7 +4,7 @@ from langchain import output_parsers
 
 from langflow.interface.base import LangChainTypeCreator
 from langflow.interface.importing.utils import import_class
-from langflow.services.getters import get_settings_service
+from langflow.services.deps import get_settings_service
 
 from langflow.template.frontend_node.output_parsers import OutputParserFrontendNode
 from loguru import logger
