+++ src/backend/langflow/services/auth/service.py
@@ -2,7 +2,7 @@ from langflow.services.base import Service
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from langflow.services.settings.manager import SettingsService
+    from langflow.services.settings.service import SettingsService
 
 
 class AuthService(Service):
