+++ src/backend/langflow/services/database/utils.py
@@ -6,7 +6,7 @@ from alembic.util.exc import CommandError
 from sqlmodel import Session
 
 if TYPE_CHECKING:
-    from langflow.services.database.manager import DatabaseService
+    from langflow.services.database.service import DatabaseService
 
 
 def initialize_database():
