+++ src/backend/langflow/services/deps.py
@@ -3,12 +3,13 @@ from typing import TYPE_CHECKING, Generator
 
 
 if TYPE_CHECKING:
-    from langflow.services.database.manager import DatabaseService
-    from langflow.services.settings.manager import SettingsService
-    from langflow.services.cache.manager import BaseCacheService
-    from langflow.services.session.manager import SessionService
-    from langflow.services.task.manager import TaskService
-    from langflow.services.chat.manager import ChatService
+    from langflow.services.database.service import DatabaseService
+    from langflow.services.settings.service import SettingsService
+    from langflow.services.cache.service import BaseCacheService
+    from langflow.services.session.service import SessionService
+    from langflow.services.task.service import TaskService
+    from langflow.services.chat.service import ChatService
+    from langflow.services.marketplace.service import MarketplaceService
     from sqlmodel import Session
 
 
@@ -46,3 +47,7 @@ def get_task_service() -> "TaskService":
 
 def get_chat_service() -> "ChatService":
     return service_manager.get(ServiceType.CHAT_SERVICE)
+
+
+def get_marketplace_service() -> "MarketplaceService":
+    return service_manager.get(ServiceType.MARKETPLACE_SERVICE)
