+++ src/backend/langflow/services/plugins/langfuse.py
@@ -1,4 +1,4 @@
-from langflow.services.getters import get_settings_service
+from langflow.services.deps import get_settings_service
 from langflow.utils.logger import logger
 
 ### Temporary implementation
