+++ src/backend/langflow/services/settings/base.py
@@ -52,6 +52,8 @@ class Settings(BaseSettings):
     LANGFUSE_PUBLIC_KEY: Optional[str] = None
     LANGFUSE_HOST: Optional[str] = None
 
+    MARKETPLACE_URL: Optional[str] = None
+
     @validator("CONFIG_DIR", pre=True, allow_reuse=True)
     def set_langflow_dir(cls, value):
         if not value:
