+++ tests/test_cli.py
@@ -3,7 +3,7 @@ from tempfile import tempdir
 from langflow.__main__ import app
 import pytest
 
-from langflow.services import getters
+from langflow.services import deps
 
 
 @pytest.fixture(scope="module")
@@ -26,7 +26,7 @@ def test_components_path(runner, client, default_settings):
         ["run", "--components-path", str(temp_dir), *default_settings],
     )
     assert result.exit_code == 0, result.stdout
-    settings_service = getters.get_settings_service()
+    settings_service = deps.get_settings_service()
     assert str(temp_dir) in settings_service.settings.COMPONENTS_PATH
