+++ tests/test_process.py
@@ -1,5 +1,5 @@
 from langflow.processing.process import process_tweaks
-from langflow.services.getters import get_session_service
+from langflow.services.deps import get_session_service
 
 
 def test_no_tweaks():
