+++ tests/test_setup_superuser.py
@@ -9,9 +9,9 @@ from langflow.services.utils import (
 )
 
 
-# @patch("langflow.services.getters.get_session")
+# @patch("langflow.services.deps.get_session")
 # @patch("langflow.services.utils.create_super_user")
-# @patch("langflow.services.getters.get_settings_service")
+# @patch("langflow.services.deps.get_settings_service")
 # # @patch("langflow.services.utils.verify_password")
 # def test_setup_superuser(
 #     mock_get_session, mock_create_super_user, mock_get_settings_service
@@ -92,8 +92,8 @@ from langflow.services.utils import (
 #     assert str(actual_expr) == str(expected_expr)
 
 
-@patch("langflow.services.getters.get_settings_service")
-@patch("langflow.services.getters.get_session")
+@patch("langflow.services.deps.get_settings_service")
+@patch("langflow.services.deps.get_session")
 def test_teardown_superuser_default_superuser(
     mock_get_session, mock_get_settings_service
 ):
@@ -120,8 +120,8 @@ def test_teardown_superuser_default_superuser(
     mock_session.commit.assert_called_once()
 
 
-@patch("langflow.services.getters.get_settings_service")
-@patch("langflow.services.getters.get_session")
+@patch("langflow.services.deps.get_settings_service")
+@patch("langflow.services.deps.get_session")
 def test_teardown_superuser_no_default_superuser(
     mock_get_session, mock_get_settings_service
 ):
